requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
Werkzeug==2.3.3
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.6
SQLAlchemy==2.0.20
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3